// types.ts
export enum Role {
  CITIZEN = 'CITIZEN',
  ADMIN = 'ADMIN',
  SUPER_ADMIN = 'SUPER_ADMIN',
  ENCODER = 'ENCODER',
  APPROVER = 'APPROVER',
  MERCHANT = 'MERCHANT',
}